+++ SIDI/steps/_react_native_npm_library_install/bitrise/config.ts
@@ -1,0 +1,7 @@
+import { IConfig } from '../../../../src/models/SidiModel';
+
+const config: IConfig = {
+  mandatoryKeys: [{ key: 'SAMPLE_APP_PATH', globalValue: true, defaultValue: 'example' }],
+};
+
+export default config;